One file to rule them all! Move inlinequeryresults to single file
Also clean up their docstrings a lot
