--- telegram/inline/inlinequeryresultgif.py
@@ -1,100 +1,0 @@
-#!/usr/bin/env python
-#
-# A library that provides a Python interface to the Telegram Bot API
-# Copyright (C) 2015-2017
-# Leandro Toledo de Souza <devs@python-telegram-bot.org>
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser Public License for more details.
-#
-# You should have received a copy of the GNU Lesser Public License
-# along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the classes that represent Telegram
-InlineQueryResultGif"""
-
-from telegram import InlineQueryResult, InlineKeyboardMarkup, InputMessageContent
-
-
-class InlineQueryResultGif(InlineQueryResult):
-    """Represents a link to an animated GIF file. By default, this animated GIF file will be sent
-    by the user with optional caption. Alternatively, you can use input_message_content to send a
-    message with the specified content instead of the animation.
-
-    Attributes:
-        gif_url (str): A valid URL for the GIF file. File size must not exceed 1MB.
-        thumb_url (str): URL of the static thumbnail for the result (jpeg or gif).
-        gif_width (Optional[int]): Width of the GIF.
-        gif_height (Optional[int]): Height of the GIF.
-        gif_duration (Optional[int]): Duration of the GIF.
-        title (Optional[str]): Title for the result.
-        caption	(Optional[str]): Caption of the GIF file to be sent, 0-200 characters.
-        reply_markup (Optional[:class:`telegram.InlineKeyboardMarkup`]): Inline keyboard attached
-            to the message.
-        input_message_content (Optional[:class:`telegram.InputMessageContent`]): Content of the
-            message to be sent instead of the GIF animation.
-
-    Args:
-        id (str):
-        gif_url (str):
-        thumb_url (str):
-        gif_width (Optional[int]):
-        gif_height (Optional[int]):
-        gif_duration (Optional[int]):
-        title (Optional[str]):
-        caption	(Optional[str]):
-        reply_markup (Optional[:class:`telegram.InlineKeyboardMarkup`]):
-        input_message_content (Optional[:class:`telegram.InputMessageContent`]):
-        **kwargs (dict): Arbitrary keyword arguments.
-
-    """
-
-    def __init__(self,
-                 id,
-                 gif_url,
-                 thumb_url,
-                 gif_width=None,
-                 gif_height=None,
-                 title=None,
-                 caption=None,
-                 reply_markup=None,
-                 input_message_content=None,
-                 gif_duration=None,
-                 **kwargs):
-
-        # Required
-        super(InlineQueryResultGif, self).__init__('gif', id)
-        self.gif_url = gif_url
-        self.thumb_url = thumb_url
-
-        # Optionals
-        if gif_width:
-            self.gif_width = gif_width
-        if gif_height:
-            self.gif_height = gif_height
-        if gif_duration:
-            self.gif_duration = gif_duration
-        if title:
-            self.title = title
-        if caption:
-            self.caption = caption
-        if reply_markup:
-            self.reply_markup = reply_markup
-        if input_message_content:
-            self.input_message_content = input_message_content
-
-    @staticmethod
-    def de_json(data, bot):
-        data = super(InlineQueryResultGif, InlineQueryResultGif).de_json(data, bot)
-
-        data['reply_markup'] = InlineKeyboardMarkup.de_json(data.get('reply_markup'), bot)
-        data['input_message_content'] = InputMessageContent.de_json(
-            data.get('input_message_content'), bot)
-
-        return InlineQueryResultGif(**data)
