--- telegram/inline/inlinequeryresultvoice.py
@@ -1,60 +1,0 @@
-#!/usr/bin/env python
-#
-# A library that provides a Python interface to the Telegram Bot API
-# Copyright (C) 2015-2017
-# Leandro Toledo de Souza <devs@python-telegram-bot.org>
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser Public License for more details.
-#
-# You should have received a copy of the GNU Lesser Public License
-# along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains the classes that represent Telegram
-InlineQueryResultVoice"""
-
-from telegram import InlineQueryResult, InlineKeyboardMarkup, InputMessageContent
-
-
-class InlineQueryResultVoice(InlineQueryResult):
-
-    def __init__(self,
-                 id,
-                 voice_url,
-                 title,
-                 voice_duration=None,
-                 caption=None,
-                 reply_markup=None,
-                 input_message_content=None,
-                 **kwargs):
-
-        # Required
-        super(InlineQueryResultVoice, self).__init__('voice', id)
-        self.voice_url = voice_url
-        self.title = title
-
-        # Optional
-        if voice_duration:
-            self.voice_duration = voice_duration
-        if caption:
-            self.caption = caption
-        if reply_markup:
-            self.reply_markup = reply_markup
-        if input_message_content:
-            self.input_message_content = input_message_content
-
-    @staticmethod
-    def de_json(data, bot):
-        data = super(InlineQueryResultVoice, InlineQueryResultVoice).de_json(data, bot)
-
-        data['reply_markup'] = InlineKeyboardMarkup.de_json(data.get('reply_markup'), bot)
-        data['input_message_content'] = InputMessageContent.de_json(
-            data.get('input_message_content'), bot)
-
-        return InlineQueryResultVoice(**data)
